rename utils, create a streaming LM pretrain
The streaming pretrainer defaults to reddit, and
similar params to convert model
